Extract parseCode into its own file
